+++ src/main/java/org/application/Main.java
@@ -1,0 +1,10 @@
+package org.application;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello, World!");
+    }
+}
