+++ src/main/java/org/application/entity/Course.java
@@ -1,0 +1,55 @@
+package org.application.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.application.enums.CoursesCategories;
+
+
+import java.time.LocalDate;
+import java.util.List;
+import java.util.Objects;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Entity
+public class Course {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+    private String title;
+    private String description;
+    LocalDate creationDate;
+    private String author;
+    LocalDate daeOfChange;
+    private String authorOfChange;
+    LocalDate deletionDate;
+    private String deletionAuthor;
+    private int rate;
+
+    private int duration;
+    private List<Module> modules;
+    private String tag;
+    private CoursesCategories category;
+
+    //*************************************
+
+    @Override
+    public boolean equals(Object o) {
+        if (o == null || getClass() != o.getClass()) return false;
+        Course course = (Course) o;
+        return id == course.id && Objects.equals(title, course.title);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, title);
+    }
+}
