+++ src/main/java/org/application/enums/AssessmentOfCourse.java
@@ -1,0 +1,14 @@
+package org.application.enums;
+
+public enum AssessmentOfCourse {
+    ONE(1),
+    TWO(2),
+    THREE(3),
+    FOUR(4),
+    FIVE(5);
+    private int value;
+
+    AssessmentOfCourse(int value) {
+        this.value = value;
+    }
+}
