+++ src/main/java/org/application/enums/Group.java
@@ -1,0 +1,4 @@
+package org.application.enums;
+
+public enum Group {
+}
