+++ src/main/java/org/application/enums/Role.java
@@ -1,0 +1,8 @@
+package org.application.enums;
+
+public enum Role {
+    OWNER,
+    ADMIN,
+    TEACHER,
+    STUDENT;
+}
