+++ src/main/java/org/application/enums/TaskType.java
@@ -1,0 +1,6 @@
+package org.application.enums;
+
+public enum TaskType {
+    TEST,
+    OPEN_ANSWERS;
+}
