+++ src/main/java/org/application/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package org.application.repository;
+
+import org.application.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Integer> {
+    User findByUsername(String username);
+
+    User findUserById(int id);
+}
