+++ src/main/java/org/application/service/RegisterServiceImpl.java
@@ -1,0 +1,23 @@
+package org.application.service;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.application.dto.UserRegisterDTO;
+import org.application.repository.UserRepository;
+import org.springframework.stereotype.Service;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class RegisterServiceImpl implements RegisterService {
+
+    private final UserRepository userRepository;
+
+    public boolean registerUser(UserRegisterDTO userRegisterDTO) {
+        if (userRepository.findByUsername(userRegisterDTO.getName()) != null) {
+            return false;
+        }
+        userService.saveRegisterUser(register);
+        return true;
+    }
+}
