+++ src/main/java/org/application/service/UserService.java
@@ -1,0 +1,4 @@
+package org.application.service;
+
+public interface UserService {
+}
