+++ src/main/java/org/application/service/UserServiceImpl.java
@@ -1,0 +1,4 @@
+package org.application.service;
+
+public class UserServiceImpl implements UserService {
+}
