+++ src/main/java/org/application/service/mapperFactoring/AddUserDTOFactoryImpl.java
@@ -1,0 +1,20 @@
+package org.application.service.mapperFactoring;
+
+import lombok.RequiredArgsConstructor;
+import org.application.dto.UserRegisterDTO;
+import org.application.entity.User;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class AddUserDTOFactoryImpl implements UserDTOFactoy {
+    /**
+     * Создание нового юзера
+     * @param newUser
+     * @return User
+     */
+    public User addNewUserFromDTO(UserRegisterDTO newUser){
+        return new User (newUser.getLogin(), newUser.getPassword(), newUser.getPassword(),
+                newUser.getPhone(), newUser.getName(), newUser.getSurname());
+    }
+}
