+++ src/main/java/org/application/service/mapperFactoring/UserDTOFactoy.java
@@ -1,0 +1,4 @@
+package org.application.service.mapperFactoring;
+
+public interface UserDTOFactoy {
+}
