--- src/main/java/org/example/Main.java
@@ -1,10 +1,0 @@
-package org.example;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello, World!");
-    }
-}
