--- src/main/java/org/example/controllers/UserController.java
@@ -1,20 +1,0 @@
-package org.example.controllers;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@Slf4j //  добавляет логгер в класс
-@CrossOrigin(value = "http://localhost:3000") // позволяет настроить CORS (Cross-Origin Resource Sharing)
-// для данного контроллера. Указывает, что этот контроллер может обрабатывать запросы с указанного домена
-// (http://localhost:3000), даже если он отличается от домена, на котором запущено приложение.
-@RequiredArgsConstructor // генерирует конструктор с аргументами для всех полей, помеченных аннотацией @NonNull
-@RestController
-@RequestMapping("/users")
-@Tag(name = "Пользователи")
-public class UserController {
-
-}
