--- src/main/java/org/example/enums/CoursesCategories.java
@@ -1,13 +1,0 @@
-package org.example.enums;
-
-public enum CoursesCategories {
-    PROGRAMMING ("Программирование"),
-    TESTING ("Тестирование"),
-    ANALYTICS("Аналитика");
-
-    private final String name;
-
-    CoursesCategories(String name) {
-        this.name = name;
-    }
-}
