--- src/main/java/org/example/enums/Group.java
@@ -1,4 +1,0 @@
-package org.example.enums;
-
-public enum Group {
-}
