--- src/main/java/org/example/enums/Role.java
@@ -1,8 +1,0 @@
-package org.example.enums;
-
-public enum Role {
-    OWNER,
-    ADMIN,
-    TEACHER,
-    STUDENT;
-}
