--- src/main/java/org/example/enums/TaskType.java
@@ -1,6 +1,0 @@
-package org.example.enums;
-
-public enum TaskType {
-    TEST,
-    OPEN_ANSWERS;
-}
