--- src/main/java/org/example/service/RegisterServiceImpl.java
@@ -1,7 +1,0 @@
-package org.example.service;
-
-import org.example.dto.UserRegisterDTO;
-
-public class RegisterServiceImpl implements RegisterService {
-    public boolean registerUser(UserRegisterDTO userRegisterDTO) {}
-}
