--- src/main/java/org/example/service/UserService.java
@@ -1,4 +1,0 @@
-package org.example.service;
-
-public interface UserService {
-}
