--- src/main/java/org/example/service/UserServiceImpl.java
@@ -1,4 +1,0 @@
-package org.example.service;
-
-public class UserServiceImpl implements UserService {
-}
