--- src/main/java/org/example/service/mapperFactoring/UserDTOFactoy.java
@@ -1,4 +1,0 @@
-package org.example.service.mapperFactoring;
-
-public interface UserDTOFactoy {
-}
